[KRACK attacks fix patch]: mac80211: accept key reinstall without changing anything.
When a key is reinstalled we can reset the replay counters
etc. which can lead to nonce reuse and/or replay detection
being impossible, breaking security properties, as described
in the "KRACK attacks".

In case this happens, simply silently accept the new key
coming from userspace but don't take any action on it since
it's the same key; this keeps the PN replay counters intact.

Signed-off-by: vijayakumar Durai <vijayakumar.durai1@vivint.com>